Fix type hints and casting in various modules for improved type safety

+++ shared/config.py
@@ -407,15 +407,15 @@ class ConfigManager:
         if os.getenv("MAX_POSITION_SIZE"):
             max_position_size = os.getenv("MAX_POSITION_SIZE")
             if max_position_size is not None:
-                config.risk.max_position_size = float(max_position_size)
+                config.risk.max_position_size = Decimal(max_position_size)
         if os.getenv("MAX_DAILY_LOSS"):
             max_daily_loss = os.getenv("MAX_DAILY_LOSS")
             if max_daily_loss is not None:
-                config.risk.max_daily_loss = float(max_daily_loss)
+                config.risk.max_daily_loss = Decimal(max_daily_loss)
         if os.getenv("STOP_LOSS_PCT"):
             stop_loss_pct = os.getenv("STOP_LOSS_PCT")
             if stop_loss_pct is not None:
-                config.risk.stop_loss_pct = float(stop_loss_pct)
+                config.risk.stop_loss_pct = Decimal(stop_loss_pct)
 
     def save_config(self, config: SyntraConfig, path: Optional[str] = None) -> None:
         """Сохранение конфигурации в файл."""
